fixing maps activity

+++ app/src/main/assets/sugar_upgrades/7.sql
@@ -1,1 +1,1 @@
-alter table USER add CLIENT_KEY LONG;
+alter table USER add CLIENT_KEY INTEGER;
